\name{Blood_specific_gene}
\alias{Blood_specific_gene}
\docType{data}
\title{Blood marker genes}
\description{Blood_specific_gene contains marker genes for Blood}
\usage{data("Blood_specific_gene")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\examples{
data(Blood_specific_gene)
## maybe str(Blood_specific_gene) ; plot(Blood_specific_gene) ...
}
\keyword{datasets}
