\name{Breast_elevated_genes}
\alias{Breast_elevated_genes}
\docType{data}
\title{Genes evelevated in breast tissue}
\description{Breast_elevated_genes having genes which are elevated in
 breast tissues as compared to other tissues}
\usage{data("Breast_elevated_genes")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}

\examples{
data(Breast_elevated_genes)
## maybe str(Breast_elevated_genes) ; plot(Breast_elevated_genes) ...
}
\keyword{datasets}
