\name{c2.all.v7.2.symbols}
\alias{c2.all.v7.2.symbols}
\docType{data}
\title{msidb c2 genesets}
\description{cr MsiDB This collection contains gene sets selected from
multiple sources, including online pathway databases and biomedical
literature. Many sets are also contributed by specific domain
specialists.}

\usage{data("c2.all.v7.2.symbols")}
\format{
  The format is:
List of 6229
 $ BIOCARTA_FEEDER_PATHWAY                                                                                                                      : chr [1:9] "MPI" "HK1" "KHK" "PGM1" ...
 $ BIOCARTA_PROTEASOME_PATHWAY                                                                                                                  : chr [1:19] "RPN1" "PSMA1" "PSMA5" "RPN2" ...
 $ BIOCARTA_KREB_PATHWAY                                                                                                                        : chr [1:8] "FH" "SDHA" "OGDH" "SUCLA2" ...
 $ ST_INTERFERON_GAMMA_PATHWAY                                                                                                                  : chr [1:10] "STAT1" "PLA2G2A" "IFNG" "JAK2" ...
 $ ST_WNT_CA2_CYCLIC_GMP_PATHWAY                                                                                                                : chr [1:20] "ITPR1" "PDE6H" "CAMK2D" "PDE6D" ...
 $ ST_DIFFERENTIATION_PATHWAY_IN_PC12_CELLS                                                                                                     : chr [1:47] "EGR3" "PTPN11" "NTRK1" "TUBA1A" ...
 $ ST_TUMOR_NECROSIS_FACTOR_PATHWAY                                                                                                             : chr [1:29] "NFKBIA" "MAP3K7" "TNFRSF1A" "MAP3K3" ...
 $ ST_ERK1_ERK2_MAPK_PATHWAY                                                                                                                    : chr [1:32] "SOS1" "RPS6KA3" "TRAF3" "CREB3" ...
 $ ST_GA12_PATHWAY                                                                                                                              : chr [1:23] "PLD1" "RAF1" "EPHB2" "VAV1" ...
 $ ST_G_ALPHA_S_PATHWAY                                                                                                                         : chr [1:16] "RAF1" "RAPGEF3" "CAMP" "BRAF" ...
 $ ST_G_ALPHA_I_PATHWAY                                                                                                                         : chr [1:36] "RAP1GAP" "EGFR" "PLCB2" "SOS1" ...
 $ ST_IL_13_PATHWAY                                                                                                                             : chr [1:7] "IL13" "IL13RA2" "JAK2" "IL13RA1" ...
 $ ST_P38_MAPK_PATHWAY                                                                                                                          : chr [1:37] "MAP3K10" "MAP3K7" "MAPK14" "CREB3" ...
 $ ST_JAK_STAT_PATHWAY                                                                                                                          : chr [1:9] "PIAS1" "JAK3" "JAK2" "CISH" ...
 $ ST_GRANULE_CELL_SURVIVAL_PATHWAY                                                                                                             : chr [1:28] "CXCR2" "CREB3" "ADCYAP1" "MAPK8IP2" ...
 $ ST_ADRENERGIC                                                                                                                                : chr [1:37] "GNA15" "EGFR" "MAPK14" "AKT1" ...
 $ ST_INTEGRIN_SIGNALING_PATHWAY                                                                                                                : chr [1:82] "ANGPTL2" "PAK3" "ARHGAP35" "ITGA4" ...
 $ ST_GAQ_PATHWAY                                                                                                                               : chr [1:28] "NFKBIA" "AKT1" "VN1R1" "NFKBIB" ...
 $ ST_GA13_PATHWAY                                                                                                                              : chr [1:37] "RDX" "NFKBIA" "AKT1" "TBXA2R" ...
 $ ST_STAT3_PATHWAY                                                                                                                             : chr [1:11] "JAK3" "JAK2" "CISH" "SRC" ...
 $ ST_T_CELL_SIGNAL_TRANSDUCTION                                                                                                                : chr [1:45] "PAK4" "LAT" "ZAP70" "VAV1" ...
 $ ST_TYPE_I_INTERFERON_PATHWAY                                                                                                                 : chr [1:9] "STAT1" "IRF9" "STAT2" "REG1A" ...
 $ ST_PAC1_RECEPTOR_PATHWAY                                                                                                                     : chr [1:8] "GNAQ" "ITPKA" "CAMP" "ADCYAP1" ...
 $ ST_B_CELL_ANTIGEN_RECEPTOR                                                                                                                   : chr [1:40] "CD19" "PPP1R13B" "VAV1" "NFKBIA" ...
 $ ST_INTERLEUKIN_4_PATHWAY                                                                                                                     : chr [1:26] "PPP1R13B" "INPP5D" "JAK2" "SOS1" ...
 $ ST_WNT_BETA_CATENIN_PATHWAY                                                                                                                  : chr [1:34] "PSEN1" "CER1" "DKK4" "NKD1" ...
 $ ST_JNK_MAPK_PATHWAY                                                                                                                          : chr [1:40] "DUSP8" "MAP3K10" "PAPPA" "MAP3K7" ...
 $ ST_FAS_SIGNALING_PATHWAY                                                                                                                     : chr [1:66] "TUFM" "EZR" "SMPD1" "MET" ...
 $ ST_MYOCYTE_AD_PATHWAY                                                                                                                        : chr [1:28] "RAC1" "RHO" "AKT1" "ITPR2" ...
 $ ST_PHOSPHOINOSITIDE_3_KINASE_PATHWAY                                                                                                         : chr [1:38] "A1BG" "YWHAZ" "PPP1R13B" "INPP5D" ...
 $ REACTOME_REMOVAL_OF_THE_FLAP_INTERMEDIATE_FROM_THE_C_STRAND                                                                                  : chr [1:11] "POLD1" "POLD3" "RPA3" "POLD2" ...
 $ BLANCO_MELO_SARS_COV_1_INFECTION_MCR5_CELLS_UP                                                                                               : chr [1:12] "CENPE" "ASPM" "ESM1" "KIF20B" ...
 $ BLANCO_MELO_SARS_COV_1_INFECTION_MCR5_CELLS_DN                                                                                               : chr [1:9] "RGCC" "FGFR4" "BMP4" "INHBB" ...
 $ BLANCO_MELO_MERS_COV_INFECTION_MCR5_CELLS_UP                                                                                                 : chr [1:295] "HSPA6" "HSPA7" "HSPA1B" "HSPA1A" ...
 $ BLANCO_MELO_MERS_COV_INFECTION_MCR5_CELLS_DN                                                                                                 : chr [1:37] "ATP5F1D" "DNAJC12" "PTGDS" "CHCHD2" ...
 $ BLANCO_MELO_INFLUENZA_A_INFECTION_A594_CELLS_UP                                                                                              : chr [1:41] "CCL5" "PYGM" "HSPA6" "LINC00641" ...
 $ BLANCO_MELO_INFLUENZA_A_INFECTION_A594_CELLS_DN                                                                                              : chr [1:118] "MAFG" "WDR18" "COL4A4" "BAX" ...
 $ BLANCO_MELO_HUMAN_PARAINFLUENZA_VIRUS_3_INFECTION_A594_CELLS_UP                                                                              : chr [1:204] "IFIT2" "IFNL1" "OASL" "IFIT1" ...
 $ BLANCO_MELO_HUMAN_PARAINFLUENZA_VIRUS_3_INFECTION_A594_CELLS_DN                                                                              : chr [1:122] "H1-10-AS1" "ADM5" "CHTF18" "TIGD3" ...
 $ BLANCO_MELO_RESPIRATORY_SYNCYTIAL_VIRUS_INFECTION_A594_CELLS_UP                                                                              : chr [1:285] "OASL" "IFNL1" "IFIT2" "IFIT1" ...
 $ BLANCO_MELO_RESPIRATORY_SYNCYTIAL_VIRUS_INFECTION_A594_CELLS_DN                                                                              : chr [1:119] "ITGA11" "CDCA7L" "MCF2L" "CCER2" ...
 $ BLANCO_MELO_COVID19_SARS_COV_2_INFECTION_A594_CELLS_UP                                                                                       : chr [1:86] "INHBE" "ADM2" "STC2" "IL1A" ...
 $ BLANCO_MELO_COVID19_SARS_COV_2_INFECTION_A594_CELLS_DN                                                                                       : chr [1:77] "HNF1A-AS1" "SPEF1" "SDF2L1" "RADIL" ...
 $ BLANCO_MELO_COVID19_SARS_COV_2_INFECTION_CALU3_CELLS_UP                                                                                      : chr [1:338] "IFNB1" "IFNL2" "IFNL3" "IFNL1" ...
 $ BLANCO_MELO_COVID19_SARS_COV_2_INFECTION_CALU3_CELLS_DN                                                                                      : chr [1:24] "MIR31HG" "TEX15" "H2AX" "SNRNP25" ...
 $ BLANCO_MELO_COVID19_SARS_COV_2_LOW_MOI_INFECTION_A594_ACE2_EXPRESSING_CELLS_UP                                                               : chr [1:174] "EGR1" "CCL20" "CXCL2" "CXCL8" ...
 $ BLANCO_MELO_COVID19_SARS_COV_2_LOW_MOI_INFECTION_A594_ACE2_EXPRESSING_CELLS_DN                                                               : chr [1:14] "CPM" "ALDH1L2" "DPP4" "KRT4" ...
 $ BLANCO_MELO_COVID19_SARS_COV_2_INFECTION_A594_ACE2_EXPRESSING_CELLS_UP                                                                       : chr [1:542] "EGR1" "CMPK2" "IFNB1" "IFI44" ...
 $ BLANCO_MELO_COVID19_SARS_COV_2_INFECTION_A594_ACE2_EXPRESSING_CELLS_DN                                                                       : chr [1:96] "SEMA3E" "FLRT2" "AKR1C4" "ANXA2" ...
 $ BLANCO_MELO_COVID19_SARS_COV_2_INFECTION_A594_ACE2_EXPRESSING_CELLS_RUXOLITINIB_UP                                                           : chr [1:483] "EGR1" "BHLHE41" "NR1D1" "PPP4R4" ...
 $ BLANCO_MELO_COVID19_SARS_COV_2_INFECTION_A594_ACE2_EXPRESSING_CELLS_RUXOLITINIB_DN                                                           : chr [1:76] "GPX2" "ALDH1A1" "ASB9" "ETFB" ...
 $ BLANCO_MELO_COVID19_BRONCHIAL_EPITHELIAL_CELLS_SARS_COV_2_INFECTION_UP                                                                       : chr [1:144] "CSF3" "CCL20" "CXCL5" "SPRR2D" ...
 $ BLANCO_MELO_COVID19_BRONCHIAL_EPITHELIAL_CELLS_SARS_COV_2_INFECTION_DN                                                                       : chr [1:58] "PDE5A" "SPTLC3" "ALDH3A1" "MN1" ...
 $ BLANCO_MELO_BETA_INTERFERON_TREATED_BRONCHIAL_EPITHELIAL_CELLS_UP                                                                            : chr [1:378] "IFIT3" "IFIT1" "IFIT2" "MX1" ...
 $ BLANCO_MELO_BETA_INTERFERON_TREATED_BRONCHIAL_EPITHELIAL_CELLS_DN                                                                            : chr [1:216] "CSTA" "MZF1" "IFITM10" "CCDC88B" ...
 $ BLANCO_MELO_BRONCHIAL_EPITHELIAL_CELLS_INFLUENZA_A_INFECTION_UP                                                                              : chr [1:136] "ITGA7" "HERC5" "PYGM" "XAF1" ...
 $ BLANCO_MELO_BRONCHIAL_EPITHELIAL_CELLS_INFLUENZA_A_INFECTION_DN                                                                              : chr [1:81] "ANKRD23" "MND1" "RECQL4" "CCDC34" ...
 $ BLANCO_MELO_BRONCHIAL_EPITHELIAL_CELLS_INFLUENZA_A_DEL_NS1_INFECTION_UP                                                                      : chr [1:644] "IFIT2" "IFIT1" "RSAD2" "CXCL10" ...
 $ BLANCO_MELO_BRONCHIAL_EPITHELIAL_CELLS_INFLUENZA_A_DEL_NS1_INFECTION_DN                                                                      : chr [1:187] "RHNO1" "RAD51" "CDCA7" "CES3" ...
 $ BLANCO_MELO_COVID19_SARS_COV_2_POS_PATIENT_LUNG_TISSUE_UP                                                                                    : chr [1:172] "FDCSP" "RNASE2" "SELL" "CYP19A1" ...
 $ BLANCO_MELO_COVID19_SARS_COV_2_POS_PATIENT_LUNG_TISSUE_DN                                                                                    : chr [1:175] "SFN" "DPH7" "THAP2" "SPSB3" ...
 $ BUFFA_HYPOXIA_METAGENE                                                                                                                       : chr [1:50] "VEGFA" "SLC2A1" "PGAM1" "ENO1" ...
 $ NAKAMURA_CANCER_MICROENVIRONMENT_UP                                                                                                          : chr [1:28] "HBA1" "PRSS2" "COL1A2" "GPIHBP1" ...
 $ NAKAMURA_CANCER_MICROENVIRONMENT_DN                                                                                                          : chr [1:46] "CDC20" "RFC4" "RFC5" "CTH" ...
 $ WEST_ADRENOCORTICAL_TUMOR_MARKERS_UP                                                                                                         : chr [1:22] "DTL" "PCSK6" "PRC1" "KIF20A" ...
 $ WEST_ADRENOCORTICAL_TUMOR_MARKERS_DN                                                                                                         : chr [1:20] "MFAP5" "IGFBP5" "KCNQ1" "RARRES2" ...
 $ WINTER_HYPOXIA_UP                                                                                                                            : chr [1:95] "HAUS2" "PTGFRN" "PYGL" "SLC6A8" ...
 $ WINTER_HYPOXIA_DN                                                                                                                            : chr [1:52] "CD48" "ANKRD44" "SYNRG" "GIMAP1" ...
 $ PARENT_MTOR_SIGNALING_UP                                                                                                                     : chr [1:568] "MSRB1" "IL17RC" "CTH" "VAT1" ...
 $ PARENT_MTOR_SIGNALING_DN                                                                                                                     : chr [1:44] "ELOVL5" "MAGEA9" "ITGAM" "KRT86" ...
 $ PYEON_HPV_POSITIVE_TUMORS_UP                                                                                                                 : chr [1:99] "PER3" "USP1" "SMARCA2" "ABCA17P" ...
 $ PYEON_HPV_POSITIVE_TUMORS_DN                                                                                                                 : chr [1:10] "NAP1L2" "PTHLH" "KLK8" "KLK10" ...
 $ NAKAMURA_TUMOR_ZONE_PERIPHERAL_VS_CENTRAL_UP                                                                                                 : chr [1:284] "MRPL42" "PNP" "NUDT5" "ETS1" ...
 $ NAKAMURA_TUMOR_ZONE_PERIPHERAL_VS_CENTRAL_DN                                                                                                 : chr [1:644] "TTC3" "SNHG7" "RAB40C" "B4GALT5" ...
 $ PICCALUGA_ANGIOIMMUNOBLASTIC_LYMPHOMA_UP                                                                                                     : chr [1:211] "EMILIN1" "KCNMA1" "CSRP2" "RPE" ...
 $ PICCALUGA_ANGIOIMMUNOBLASTIC_LYMPHOMA_DN                                                                                                     : chr [1:137] "MYLIP" "CBX4" "NR4A2" "CREM" ...
 $ LU_TUMOR_VASCULATURE_UP                                                                                                                      : chr [1:29] "WDR77" "TNFAIP6" "FJX1" "ISG15" ...
 $ LU_TUMOR_VASCULATURE_DN                                                                                                                      : chr [1:13] "ARX" "ABI3BP" "CELF2" "PRKAR2B" ...
 $ KORKOLA_EMBRYONAL_CARCINOMA_UP                                                                                                               : chr [1:42] "LRP6" "NANOG" "C12orf57" "PLBD1" ...
 $ KORKOLA_EMBRYONAL_CARCINOMA_DN                                                                                                               : chr [1:13] "SOX5" "LINC02366" "LRRC23" "H2AJ" ...
 $ KORKOLA_TERATOMA_UP                                                                                                                          : chr [1:17] "C12orf57" "CCND2" "ITFG2" "RECQL" ...
 $ KORKOLA_YOLK_SAC_TUMOR_UP                                                                                                                    : chr [1:20] "GAPDH" "DDX47" "TPI1" "SLC2A3" ...
 $ KORKOLA_CHORIOCARCINOMA_UP                                                                                                                   : chr [1:6] "CD9" "TPI1" "PLBD1" "GAPDH" ...
 $ KORKOLA_CHORIOCARCINOMA_DN                                                                                                                   : chr [1:11] "CLEC12B" "ACRBP" "CASC1" "RASSF8" ...
 $ KORKOLA_SEMINOMA_UP                                                                                                                          : chr [1:44] "TEAD4" "CD27" "DSTNP2" "ARHGDIB" ...
 $ KORKOLA_SEMINOMA_DN                                                                                                                          : chr [1:13] "LRRC23" "CLEC12B" "SOX5" "RASSF8" ...
 $ WATANABE_RECTAL_CANCER_RADIOTHERAPY_RESPONSIVE_UP                                                                                            : chr [1:113] "BTN3A1" "NELFA" "GAS6" "TRIM44" ...
 $ WATANABE_RECTAL_CANCER_RADIOTHERAPY_RESPONSIVE_DN                                                                                            : chr [1:93] "MAP4" "SCARB1" "WSB1" "SPR" ...
 $ HOLLMANN_APOPTOSIS_VIA_CD40_UP                                                                                                               : chr [1:203] "ATP8A1" "PPP3CC" "TUBA1A" "MAP4K4" ...
 $ HOLLMANN_APOPTOSIS_VIA_CD40_DN                                                                                                               : chr [1:267] "MSRB2" "CDKN1B" "HLA-DRB4" "VPS72" ...
 $ LIU_PROSTATE_CANCER_UP                                                                                                                       : chr [1:98] "BEND4" "CAMKK2" "TRIM68" "SOX4" ...
 $ LIU_PROSTATE_CANCER_DN                                                                                                                       : chr [1:491] "ITGA8" "ARHGEF40" "AKAP12" "VEGFA" ...
 $ LIU_SOX4_TARGETS_UP                                                                                                                          : chr [1:139] "ATP2B4" "INTS9" "TMSB10" "KLHL24" ...
 $ LIU_SOX4_TARGETS_DN                                                                                                                          : chr [1:310] "SHROOM3" "SURF6" "DECR1" "KIAA0319L" ...
 $ ONKEN_UVEAL_MELANOMA_UP                                                                                                                      : chr [1:790] "BAZ2B" "SNW1" "ADAM23" "CDC25B" ...
 $ ONKEN_UVEAL_MELANOMA_DN                                                                                                                      : chr [1:529] "SMARCC1" "CSNK2A2" "EIF2D" "RAB11FIP1" ...
 $ BERTUCCI_MEDULLARY_VS_DUCTAL_BREAST_CANCER_UP                                                                                                : chr [1:206] "YBX3" "CALHM6" "CDCP1" "RILPL2" ...
 $ BERTUCCI_MEDULLARY_VS_DUCTAL_BREAST_CANCER_DN                                                                                                : chr [1:180] "ZBTB43" "LSP1P4" "SNHG14" "ATL1" ...
 $ SCHUETZ_BREAST_CANCER_DUCTAL_INVASIVE_UP                                                                                                     : chr [1:352] "LPXN" "NUAK1" "IL1R1" "SPARC" ...
  [list output truncated]
}

\examples{
data(c2.all.v7.2.symbols)
## maybe str(c2.all.v7.2.symbols) ; plot(c2.all.v7.2.symbols) ...
}
\keyword{datasets}
