% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_alterations.R
\name{CNV_alterations}
\alias{CNV_alterations}
\title{CNV in p and q arms of chromosome}
\usage{
CNV_alterations(
  data_list = list(),
  data_id = list(),
  min_Sample = 5,
  min_Gene = 1500,
  path = " ",
  GenePositionFile = " ",
  threads_no = 8,
  MetaData = list(),
  Groupby = " ",
  cutoff = 1,
  Reference_name = " ",
  obs.title = "Observations",
  ref.title = "References",
  out.Filename = "inferCNV"
)
}
\arguments{
\item{data_list}{List of raw expression matrix. Genes should be in
rows and cells should be in columns in each data in the list.}

\item{data_id}{List of names/ids of expression matrix}

\item{min_Sample}{gene filter, filter out genes which are not expressed
in at least min_Sample cells}

\item{min_Gene}{cell filter, filter out those cells which do not
express at least min_Gene genes}

\item{path}{Path of output directory to save results}

\item{GenePositionFile}{A gene/chromosome positions file with chromosome
name, start, end position. "genecode hg19" positional file is given with
this package. Either you can use same using unCTC::gencode_v19_gene_pos
or can download from other sources.}

\item{threads_no}{(int) number of threads for parallel steps (default: 8)}

\item{MetaData}{Optional, List of metadata of expression matricies in same
order in which expression matricies in data_list, Column number and names
of all the MetaData in the list must be same}

\item{Groupby}{Any column name from MetaData,which we want to use as
annotation file. Only applicable if MetaData is included.}

\item{cutoff}{The minimum average read counts per gene among
reference cells. (The default value is 1)}

\item{Reference_name}{Any one type of cell from data_id list or
any one cell type from column assign to Groupby.}

\item{obs.title}{Title of test/observation matrix.
Default is "Observations"}

\item{ref.title}{Title of reference matrix. Default is "References".}

\item{out.Filename}{Store results with out.Filename prefix.
Default is "inferCNV".}
}
\value{
p_and_q_arm_CNV
}
\description{
Use single cell RNA-Seq expression data to identify copy
number variation at chromosomal level such as deletions or gains of
entire chromosome or large segments of chromosome
}
\examples{
data1 = unCTC::Poonia_et_al._TPMData
metadata = list(unCTC::Poonia_et_al._metaData)
Data_list = list(data1)
Data_Id = list("data1")
GenePoFile = unCTC::gencode_v19_gene_pos
GroupID = "GroupID"
ref = "P1"
path = getwd()
CNV_alterations(data_list=Data_list,
                data_id= Data_Id,
                min_Sample = 5,
                min_Gene = 1500,
                path=path,
                GenePositionFile=GenePoFile,
                threads_no=1,
                MetaData=metadata,
                Groupby=GroupID,
                Reference_name=ref,
                obs.title ="Observations",
                ref.title = "References",
                out.Filename = "inferCNV"
                )

}
