% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSingleCellObject.R
\name{CreateSingleCellObject}
\alias{CreateSingleCellObject}
\title{Create SingleCellExperiment instance}
\usage{
CreateSingleCellObject(data_list = list(), min_sample = 5, min_gene = 1500)
}
\arguments{
\item{data_list}{List of expression matrix.
Genes should be in rows and cells should be
in columns in each data in the list.}

\item{min_sample}{gene filter, filter out genes which are not expressed
in at least min_sample cells}

\item{min_gene}{cell filter, filter out those cells which do not
express at least min_gene genes}
}
\value{
sce SingleCellExperimemnt instance od data
 passed in a list.
}
\description{
Creating SingleCellExperiment
 instance of the data. Steps in creating
 SingleCellExperiment instance are like this.
1. Integrate all the data matrix passed in the
list on the basois of common genes.
2. Filter out low quality data.
3. Use Linnorm.Norm function to remove batch
effect between data taken from different
experiments/bathes
4. Create SingleCellExperiment instances. In
this instance two type of data format present.

Raw_filtered_data = Low quality gene and cell filtered
out matrix without batch effect correction.
}
\examples{
data1 = unCTC::Poonia_et_al._TPMData
data2 = unCTC::Ding_et_al._WBC1_TPMData
Data_list = list(data1,data2)
sce_obj = CreateSingleCellObject(data_list=Data_list,
                                  min_sample =5,
                                  min_gene=1500)

}
