% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSingleCellObject.R
\name{data_filtering}
\alias{data_filtering}
\title{Filter out poor quality data}
\usage{
data_filtering(data_mat, min_sample = 5, min_gene = 1500)
}
\arguments{
\item{data_mat}{Expression matrix, genes in row and cells in column}

\item{min_sample}{gene filter, filter out genes which are not expressed
in at least min_sample cells}

\item{min_gene}{cell filter, filter out those cells which do not
express at least min_gene genes}
}
\value{
Filtered_data return filtered data
}
\description{
Filter out poor quality genes and cells
}
\examples{
data = unCTC::Poonia_et_al._TPMData
Filtered_data = data_filtering(data_mat= data,
                               min_sample =5,
                               min_gene=1500)
}
