% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSingleCellObject.R
\name{Data_integration}
\alias{Data_integration}
\title{Integrate single cell expression matricies in the list}
\usage{
Data_integration(data_list)
}
\arguments{
\item{data_list}{List of data matrix. Genes/Features
should be in rows and cells/ samples in columns.}
}
\value{
integrated_data Return integrated data matrix
 in which genes in rows and cells in column.
}
\description{
Integrates all the data matrix present
 in the list on the basis of common genes.
}
\examples{
data1 <- unCTC::Poonia_et_al._TPMData
data2 = unCTC::Ding_et_al._WBC1_TPMData
Data_list = list(data1,data2)
Integrated_data = Data_integration(data_list=Data_list)
}
