% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDLK_Clust.R
\name{DDLK_Clust}
\alias{DDLK_Clust}
\title{Cluster pathway mapped expression mtrix using Deep Dictionary Learning
with K-means clustering cost}
\usage{
DDLK_Clust(
  PathwayScore,
  PathwayMetaData,
  n,
  out.dir = getwd(),
  MetaData = list()
)
}
\arguments{
\item{PathwayScore}{Pathway_score matrix form PathwayEnrichmentScore output}

\item{PathwayMetaData}{Pathway_metadata form PathwayEnrichmentScore output}

\item{n}{integer number of clusters for k means.}

\item{out.dir}{= Output directory to write Pathwayscore.
Default is current directory.}

\item{MetaData}{Optional, List of metadata of expression matricies in same
order in which expression matricies in data_list, Column number and names
of all the MetaData in the list must be same}
}
\value{
Pathways_score_cluster Return:
1. Pathway_score = Cell wise pathway enrichment score matrix,
   Pathways in row and cells/samples in column.
2. PathwayDDLK_clust = sample/Cell wise DDLK cluster information.
}
\description{
clustering datasets with thousands of samples, and
incorporating the K-means clustering cost into the deep dictionary
learning (DDL) framework
}
\examples{
data1 = unCTC::Poonia_et_al._TPMData
data2 = unCTC::Ding_et_al._WBC1_TPMData
Data_list = list(data1,data2)
Data_Id = list("data1","data2")
Genesets = unCTC::c2.all.v7.2.symbols
Pathway_score = PathwayEnrichmentScore(data_list=Data_list,
                                        data_id= Data_Id,
                                        Genesets=Genesets,
                                        min.size=70,
                                        max.size=100)

cluster_output = DDLK_Clust(PathwayScore = Pathway_score$Pathway_score,
                           PathwayMetaData=Pathway_score$Pathway_metadata,
                            n=3,
                            out.dir = getwd())

}
