% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupsDiffPathways.R
\name{Differential_pathways}
\alias{Differential_pathways}
\title{Calculate differential pathways between cells/clusters.}
\usage{
Differential_pathways(
  Pathway_score,
  DDLK_Clusters,
  DifferentiateBy = "Clusters",
  p_val = 0.05,
  lfc = 0,
  up_pathways_number = 10
)
}
\arguments{
\item{Pathway_score}{Output of PathwayEnrichmentScore.R method}

\item{DDLK_Clusters}{Output of DDLK_Clust.R method}

\item{DifferentiateBy}{Any column name from
DDLK_Clusters$PathwayDDLK_clust, default is "Clusters"}

\item{p_val}{Threshold p value, default is 0.05}

\item{lfc}{Threshold log fold change value, default is 0}

\item{up_pathways_number}{select number of upregulated pathways
in each group}
}
\value{
Diff_Pathways list of three objects
1. DiffMatpathway = Pathway matrix with top most
up_pathways_number differential pathways in each group
2. Diffup_pathways = top most up_pathways_number differential
 pathways in each group
3. annotations = Cell wise annotation of DifferentialMatrix
}
\description{
Provide differential pathways between given groups.
}
\examples{
data1 = unCTC::Poonia_et_al._TPMData
data2 = unCTC::Ding_et_al._WBC1_TPMData
Data_list = list(data1,data2)
Data_Id = list("data1","data2")
Genesets = unCTC::c2.all.v7.2.symbols
Pathway_score = PathwayEnrichmentScore(data_list=Data_list,
                                        data_id= Data_Id,
                                         min_Sample = 5,
                                         min_Gene = 1500,
                                        Genesets=Genesets,
                                        min.size=70,
                                        max.size=100)

DDLK_Clusters = DDLK_Clust(PathwayScore = Pathway_score$Pathway_score,
                           PathwayMetaData=Pathway_score$Pathway_metadata,
                            n=3,
                            out.dir = getwd())

Output = Differential_pathways(Pathway_score=Pathway_score,
                              DDLK_Clusters=DDLK_Clusters,
                              DifferentiateBy = "Clusters")
}
