\name{Ding_et_al._WBC1_metaData}
\alias{Ding_et_al._WBC1_metaData}
\docType{data}
\title{Ding et al. WBC1 metaData}
\description{Meta data file of single cell RNA-seq PBMC1 experiment employing
Smart-seq2 technology from the single cell broad institute's "Single Cell
Comparison: PBMC Study."}
\usage{data("Ding_et_al._WBC1_metaData")}
\format{
  A data frame with 376 observations on the following 4 variables.
  \describe{
    \item{\code{GroupID}}{a character vector}
    \item{\code{Class}}{a character vector}
    \item{\code{Cell_type}}{a character vector}
    \item{\code{HormoneStatus}}{a character vector}
  }
}

\examples{
data(Ding_et_al._WBC1_metaData)
## maybe str(Ding_et_al._WBC1_metaData) ; plot(Ding_et_al._WBC1_metaData) ...
}
\keyword{datasets}
