\name{Ding_et_al._WBC1_TPMData}
\alias{Ding_et_al._WBC1_TPMData}
\docType{data}
\title{Ding et al. WBC1 TPMData}
\description{Expression matrix (TPM) of single cell RNA-seq PBMC1 experiment
employing Smart-seq2 technology from the single cell broad institute's
"Single Cell Comparison: PBMC Study.}
\usage{data("Ding_et_al._WBC1_TPMData")}
\format{
  A data frame with 33694 observations on the following 376 variables.
  \describe{
    \item{\code{Cell_1}}{a numeric vector}
    \item{\code{Cell_10}}{a numeric vector}
    \item{\code{Cell_100}}{a numeric vector}
    \item{\code{Cell_101}}{a numeric vector}
    \item{\code{Cell_102}}{a numeric vector}
    \item{\code{Cell_103}}{a numeric vector}
    \item{\code{Cell_104}}{a numeric vector}
    \item{\code{Cell_105}}{a numeric vector}
    \item{\code{Cell_106}}{a numeric vector}
    \item{\code{Cell_107}}{a numeric vector}
    \item{\code{Cell_108}}{a numeric vector}
    \item{\code{Cell_109}}{a numeric vector}
    \item{\code{Cell_11}}{a numeric vector}
    \item{\code{Cell_110}}{a numeric vector}
    \item{\code{Cell_111}}{a numeric vector}
    \item{\code{Cell_112}}{a numeric vector}
    \item{\code{Cell_113}}{a numeric vector}
    \item{\code{Cell_114}}{a numeric vector}
    \item{\code{Cell_115}}{a numeric vector}
    \item{\code{Cell_116}}{a numeric vector}
    \item{\code{Cell_117}}{a numeric vector}
    \item{\code{Cell_118}}{a numeric vector}
    \item{\code{Cell_119}}{a numeric vector}
    \item{\code{Cell_12}}{a numeric vector}
    \item{\code{Cell_120}}{a numeric vector}
    \item{\code{Cell_121}}{a numeric vector}
    \item{\code{Cell_122}}{a numeric vector}
    \item{\code{Cell_123}}{a numeric vector}
    \item{\code{Cell_124}}{a numeric vector}
    \item{\code{Cell_125}}{a numeric vector}
    \item{\code{Cell_126}}{a numeric vector}
    \item{\code{Cell_127}}{a numeric vector}
    \item{\code{Cell_128}}{a numeric vector}
    \item{\code{Cell_129}}{a numeric vector}
    \item{\code{Cell_13}}{a numeric vector}
    \item{\code{Cell_130}}{a numeric vector}
    \item{\code{Cell_131}}{a numeric vector}
    \item{\code{Cell_132}}{a numeric vector}
    \item{\code{Cell_133}}{a numeric vector}
    \item{\code{Cell_134}}{a numeric vector}
    \item{\code{Cell_135}}{a numeric vector}
    \item{\code{Cell_136}}{a numeric vector}
    \item{\code{Cell_137}}{a numeric vector}
    \item{\code{Cell_138}}{a numeric vector}
    \item{\code{Cell_139}}{a numeric vector}
    \item{\code{Cell_14}}{a numeric vector}
    \item{\code{Cell_140}}{a numeric vector}
    \item{\code{Cell_141}}{a numeric vector}
    \item{\code{Cell_142}}{a numeric vector}
    \item{\code{Cell_143}}{a numeric vector}
    \item{\code{Cell_144}}{a numeric vector}
    \item{\code{Cell_145}}{a numeric vector}
    \item{\code{Cell_146}}{a numeric vector}
    \item{\code{Cell_147}}{a numeric vector}
    \item{\code{Cell_148}}{a numeric vector}
    \item{\code{Cell_149}}{a numeric vector}
    \item{\code{Cell_15}}{a numeric vector}
    \item{\code{Cell_150}}{a numeric vector}
    \item{\code{Cell_151}}{a numeric vector}
    \item{\code{Cell_152}}{a numeric vector}
    \item{\code{Cell_153}}{a numeric vector}
    \item{\code{Cell_154}}{a numeric vector}
    \item{\code{Cell_155}}{a numeric vector}
    \item{\code{Cell_156}}{a numeric vector}
    \item{\code{Cell_157}}{a numeric vector}
    \item{\code{Cell_158}}{a numeric vector}
    \item{\code{Cell_159}}{a numeric vector}
    \item{\code{Cell_16}}{a numeric vector}
    \item{\code{Cell_160}}{a numeric vector}
    \item{\code{Cell_161}}{a numeric vector}
    \item{\code{Cell_162}}{a numeric vector}
    \item{\code{Cell_163}}{a numeric vector}
    \item{\code{Cell_164}}{a numeric vector}
    \item{\code{Cell_165}}{a numeric vector}
    \item{\code{Cell_166}}{a numeric vector}
    \item{\code{Cell_167}}{a numeric vector}
    \item{\code{Cell_168}}{a numeric vector}
    \item{\code{Cell_169}}{a numeric vector}
    \item{\code{Cell_17}}{a numeric vector}
    \item{\code{Cell_170}}{a numeric vector}
    \item{\code{Cell_171}}{a numeric vector}
    \item{\code{Cell_173}}{a numeric vector}
    \item{\code{Cell_174}}{a numeric vector}
    \item{\code{Cell_175}}{a numeric vector}
    \item{\code{Cell_176}}{a numeric vector}
    \item{\code{Cell_177}}{a numeric vector}
    \item{\code{Cell_178}}{a numeric vector}
    \item{\code{Cell_179}}{a numeric vector}
    \item{\code{Cell_18}}{a numeric vector}
    \item{\code{Cell_180}}{a numeric vector}
    \item{\code{Cell_181}}{a numeric vector}
    \item{\code{Cell_182}}{a numeric vector}
    \item{\code{Cell_183}}{a numeric vector}
    \item{\code{Cell_185}}{a numeric vector}
    \item{\code{Cell_186}}{a numeric vector}
    \item{\code{Cell_187}}{a numeric vector}
    \item{\code{Cell_188}}{a numeric vector}
    \item{\code{Cell_189}}{a numeric vector}
    \item{\code{Cell_19}}{a numeric vector}
    \item{\code{Cell_190}}{a numeric vector}
    \item{\code{Cell_191}}{a numeric vector}
    \item{\code{Cell_192}}{a numeric vector}
    \item{\code{Cell_193}}{a numeric vector}
    \item{\code{Cell_194}}{a numeric vector}
    \item{\code{Cell_195}}{a numeric vector}
    \item{\code{Cell_196}}{a numeric vector}
    \item{\code{Cell_197}}{a numeric vector}
    \item{\code{Cell_198}}{a numeric vector}
    \item{\code{Cell_199}}{a numeric vector}
    \item{\code{Cell_2}}{a numeric vector}
    \item{\code{Cell_20}}{a numeric vector}
    \item{\code{Cell_200}}{a numeric vector}
    \item{\code{Cell_201}}{a numeric vector}
    \item{\code{Cell_202}}{a numeric vector}
    \item{\code{Cell_203}}{a numeric vector}
    \item{\code{Cell_204}}{a numeric vector}
    \item{\code{Cell_205}}{a numeric vector}
    \item{\code{Cell_206}}{a numeric vector}
    \item{\code{Cell_207}}{a numeric vector}
    \item{\code{Cell_208}}{a numeric vector}
    \item{\code{Cell_209}}{a numeric vector}
    \item{\code{Cell_21}}{a numeric vector}
    \item{\code{Cell_210}}{a numeric vector}
    \item{\code{Cell_211}}{a numeric vector}
    \item{\code{Cell_212}}{a numeric vector}
    \item{\code{Cell_213}}{a numeric vector}
    \item{\code{Cell_214}}{a numeric vector}
    \item{\code{Cell_215}}{a numeric vector}
    \item{\code{Cell_216}}{a numeric vector}
    \item{\code{Cell_217}}{a numeric vector}
    \item{\code{Cell_218}}{a numeric vector}
    \item{\code{Cell_219}}{a numeric vector}
    \item{\code{Cell_22}}{a numeric vector}
    \item{\code{Cell_220}}{a numeric vector}
    \item{\code{Cell_221}}{a numeric vector}
    \item{\code{Cell_222}}{a numeric vector}
    \item{\code{Cell_223}}{a numeric vector}
    \item{\code{Cell_224}}{a numeric vector}
    \item{\code{Cell_225}}{a numeric vector}
    \item{\code{Cell_226}}{a numeric vector}
    \item{\code{Cell_227}}{a numeric vector}
    \item{\code{Cell_228}}{a numeric vector}
    \item{\code{Cell_229}}{a numeric vector}
    \item{\code{Cell_23}}{a numeric vector}
    \item{\code{Cell_230}}{a numeric vector}
    \item{\code{Cell_231}}{a numeric vector}
    \item{\code{Cell_232}}{a numeric vector}
    \item{\code{Cell_233}}{a numeric vector}
    \item{\code{Cell_234}}{a numeric vector}
    \item{\code{Cell_235}}{a numeric vector}
    \item{\code{Cell_236}}{a numeric vector}
    \item{\code{Cell_237}}{a numeric vector}
    \item{\code{Cell_238}}{a numeric vector}
    \item{\code{Cell_239}}{a numeric vector}
    \item{\code{Cell_24}}{a numeric vector}
    \item{\code{Cell_240}}{a numeric vector}
    \item{\code{Cell_241}}{a numeric vector}
    \item{\code{Cell_242}}{a numeric vector}
    \item{\code{Cell_243}}{a numeric vector}
    \item{\code{Cell_244}}{a numeric vector}
    \item{\code{Cell_245}}{a numeric vector}
    \item{\code{Cell_246}}{a numeric vector}
    \item{\code{Cell_247}}{a numeric vector}
    \item{\code{Cell_248}}{a numeric vector}
    \item{\code{Cell_249}}{a numeric vector}
    \item{\code{Cell_25}}{a numeric vector}
    \item{\code{Cell_250}}{a numeric vector}
    \item{\code{Cell_251}}{a numeric vector}
    \item{\code{Cell_252}}{a numeric vector}
    \item{\code{Cell_253}}{a numeric vector}
    \item{\code{Cell_254}}{a numeric vector}
    \item{\code{Cell_255}}{a numeric vector}
    \item{\code{Cell_256}}{a numeric vector}
    \item{\code{Cell_257}}{a numeric vector}
    \item{\code{Cell_258}}{a numeric vector}
    \item{\code{Cell_259}}{a numeric vector}
    \item{\code{Cell_26}}{a numeric vector}
    \item{\code{Cell_260}}{a numeric vector}
    \item{\code{Cell_261}}{a numeric vector}
    \item{\code{Cell_262}}{a numeric vector}
    \item{\code{Cell_263}}{a numeric vector}
    \item{\code{Cell_264}}{a numeric vector}
    \item{\code{Cell_265}}{a numeric vector}
    \item{\code{Cell_266}}{a numeric vector}
    \item{\code{Cell_267}}{a numeric vector}
    \item{\code{Cell_269}}{a numeric vector}
    \item{\code{Cell_27}}{a numeric vector}
    \item{\code{Cell_270}}{a numeric vector}
    \item{\code{Cell_271}}{a numeric vector}
    \item{\code{Cell_272}}{a numeric vector}
    \item{\code{Cell_273}}{a numeric vector}
    \item{\code{Cell_274}}{a numeric vector}
    \item{\code{Cell_275}}{a numeric vector}
    \item{\code{Cell_276}}{a numeric vector}
    \item{\code{Cell_277}}{a numeric vector}
    \item{\code{Cell_278}}{a numeric vector}
    \item{\code{Cell_279}}{a numeric vector}
    \item{\code{Cell_28}}{a numeric vector}
    \item{\code{Cell_281}}{a numeric vector}
    \item{\code{Cell_282}}{a numeric vector}
    \item{\code{Cell_283}}{a numeric vector}
    \item{\code{Cell_284}}{a numeric vector}
    \item{\code{Cell_285}}{a numeric vector}
    \item{\code{Cell_286}}{a numeric vector}
    \item{\code{Cell_287}}{a numeric vector}
    \item{\code{Cell_288}}{a numeric vector}
    \item{\code{Cell_289}}{a numeric vector}
    \item{\code{Cell_29}}{a numeric vector}
    \item{\code{Cell_290}}{a numeric vector}
    \item{\code{Cell_291}}{a numeric vector}
    \item{\code{Cell_292}}{a numeric vector}
    \item{\code{Cell_293}}{a numeric vector}
    \item{\code{Cell_294}}{a numeric vector}
    \item{\code{Cell_295}}{a numeric vector}
    \item{\code{Cell_296}}{a numeric vector}
    \item{\code{Cell_297}}{a numeric vector}
    \item{\code{Cell_298}}{a numeric vector}
    \item{\code{Cell_299}}{a numeric vector}
    \item{\code{Cell_3}}{a numeric vector}
    \item{\code{Cell_30}}{a numeric vector}
    \item{\code{Cell_300}}{a numeric vector}
    \item{\code{Cell_301}}{a numeric vector}
    \item{\code{Cell_302}}{a numeric vector}
    \item{\code{Cell_303}}{a numeric vector}
    \item{\code{Cell_304}}{a numeric vector}
    \item{\code{Cell_305}}{a numeric vector}
    \item{\code{Cell_306}}{a numeric vector}
    \item{\code{Cell_307}}{a numeric vector}
    \item{\code{Cell_308}}{a numeric vector}
    \item{\code{Cell_309}}{a numeric vector}
    \item{\code{Cell_31}}{a numeric vector}
    \item{\code{Cell_310}}{a numeric vector}
    \item{\code{Cell_311}}{a numeric vector}
    \item{\code{Cell_312}}{a numeric vector}
    \item{\code{Cell_313}}{a numeric vector}
    \item{\code{Cell_314}}{a numeric vector}
    \item{\code{Cell_315}}{a numeric vector}
    \item{\code{Cell_316}}{a numeric vector}
    \item{\code{Cell_317}}{a numeric vector}
    \item{\code{Cell_318}}{a numeric vector}
    \item{\code{Cell_319}}{a numeric vector}
    \item{\code{Cell_32}}{a numeric vector}
    \item{\code{Cell_320}}{a numeric vector}
    \item{\code{Cell_321}}{a numeric vector}
    \item{\code{Cell_322}}{a numeric vector}
    \item{\code{Cell_323}}{a numeric vector}
    \item{\code{Cell_324}}{a numeric vector}
    \item{\code{Cell_325}}{a numeric vector}
    \item{\code{Cell_326}}{a numeric vector}
    \item{\code{Cell_327}}{a numeric vector}
    \item{\code{Cell_328}}{a numeric vector}
    \item{\code{Cell_329}}{a numeric vector}
    \item{\code{Cell_33}}{a numeric vector}
    \item{\code{Cell_330}}{a numeric vector}
    \item{\code{Cell_331}}{a numeric vector}
    \item{\code{Cell_332}}{a numeric vector}
    \item{\code{Cell_333}}{a numeric vector}
    \item{\code{Cell_334}}{a numeric vector}
    \item{\code{Cell_335}}{a numeric vector}
    \item{\code{Cell_336}}{a numeric vector}
    \item{\code{Cell_337}}{a numeric vector}
    \item{\code{Cell_338}}{a numeric vector}
    \item{\code{Cell_339}}{a numeric vector}
    \item{\code{Cell_34}}{a numeric vector}
    \item{\code{Cell_340}}{a numeric vector}
    \item{\code{Cell_341}}{a numeric vector}
    \item{\code{Cell_342}}{a numeric vector}
    \item{\code{Cell_343}}{a numeric vector}
    \item{\code{Cell_344}}{a numeric vector}
    \item{\code{Cell_345}}{a numeric vector}
    \item{\code{Cell_346}}{a numeric vector}
    \item{\code{Cell_347}}{a numeric vector}
    \item{\code{Cell_348}}{a numeric vector}
    \item{\code{Cell_349}}{a numeric vector}
    \item{\code{Cell_35}}{a numeric vector}
    \item{\code{Cell_350}}{a numeric vector}
    \item{\code{Cell_351}}{a numeric vector}
    \item{\code{Cell_352}}{a numeric vector}
    \item{\code{Cell_353}}{a numeric vector}
    \item{\code{Cell_354}}{a numeric vector}
    \item{\code{Cell_355}}{a numeric vector}
    \item{\code{Cell_356}}{a numeric vector}
    \item{\code{Cell_357}}{a numeric vector}
    \item{\code{Cell_358}}{a numeric vector}
    \item{\code{Cell_359}}{a numeric vector}
    \item{\code{Cell_36}}{a numeric vector}
    \item{\code{Cell_360}}{a numeric vector}
    \item{\code{Cell_361}}{a numeric vector}
    \item{\code{Cell_362}}{a numeric vector}
    \item{\code{Cell_363}}{a numeric vector}
    \item{\code{Cell_365}}{a numeric vector}
    \item{\code{Cell_366}}{a numeric vector}
    \item{\code{Cell_367}}{a numeric vector}
    \item{\code{Cell_368}}{a numeric vector}
    \item{\code{Cell_369}}{a numeric vector}
    \item{\code{Cell_37}}{a numeric vector}
    \item{\code{Cell_370}}{a numeric vector}
    \item{\code{Cell_371}}{a numeric vector}
    \item{\code{Cell_372}}{a numeric vector}
    \item{\code{Cell_373}}{a numeric vector}
    \item{\code{Cell_374}}{a numeric vector}
    \item{\code{Cell_375}}{a numeric vector}
    \item{\code{Cell_377}}{a numeric vector}
    \item{\code{Cell_378}}{a numeric vector}
    \item{\code{Cell_379}}{a numeric vector}
    \item{\code{Cell_38}}{a numeric vector}
    \item{\code{Cell_380}}{a numeric vector}
    \item{\code{Cell_381}}{a numeric vector}
    \item{\code{Cell_382}}{a numeric vector}
    \item{\code{Cell_383}}{a numeric vector}
    \item{\code{Cell_384}}{a numeric vector}
    \item{\code{Cell_39}}{a numeric vector}
    \item{\code{Cell_4}}{a numeric vector}
    \item{\code{Cell_40}}{a numeric vector}
    \item{\code{Cell_41}}{a numeric vector}
    \item{\code{Cell_43}}{a numeric vector}
    \item{\code{Cell_44}}{a numeric vector}
    \item{\code{Cell_45}}{a numeric vector}
    \item{\code{Cell_46}}{a numeric vector}
    \item{\code{Cell_47}}{a numeric vector}
    \item{\code{Cell_48}}{a numeric vector}
    \item{\code{Cell_49}}{a numeric vector}
    \item{\code{Cell_5}}{a numeric vector}
    \item{\code{Cell_50}}{a numeric vector}
    \item{\code{Cell_51}}{a numeric vector}
    \item{\code{Cell_52}}{a numeric vector}
    \item{\code{Cell_53}}{a numeric vector}
    \item{\code{Cell_54}}{a numeric vector}
    \item{\code{Cell_55}}{a numeric vector}
    \item{\code{Cell_56}}{a numeric vector}
    \item{\code{Cell_57}}{a numeric vector}
    \item{\code{Cell_58}}{a numeric vector}
    \item{\code{Cell_59}}{a numeric vector}
    \item{\code{Cell_6}}{a numeric vector}
    \item{\code{Cell_60}}{a numeric vector}
    \item{\code{Cell_61}}{a numeric vector}
    \item{\code{Cell_62}}{a numeric vector}
    \item{\code{Cell_63}}{a numeric vector}
    \item{\code{Cell_64}}{a numeric vector}
    \item{\code{Cell_65}}{a numeric vector}
    \item{\code{Cell_66}}{a numeric vector}
    \item{\code{Cell_67}}{a numeric vector}
    \item{\code{Cell_68}}{a numeric vector}
    \item{\code{Cell_69}}{a numeric vector}
    \item{\code{Cell_7}}{a numeric vector}
    \item{\code{Cell_70}}{a numeric vector}
    \item{\code{Cell_71}}{a numeric vector}
    \item{\code{Cell_72}}{a numeric vector}
    \item{\code{Cell_73}}{a numeric vector}
    \item{\code{Cell_74}}{a numeric vector}
    \item{\code{Cell_75}}{a numeric vector}
    \item{\code{Cell_76}}{a numeric vector}
    \item{\code{Cell_77}}{a numeric vector}
    \item{\code{Cell_78}}{a numeric vector}
    \item{\code{Cell_79}}{a numeric vector}
    \item{\code{Cell_8}}{a numeric vector}
    \item{\code{Cell_80}}{a numeric vector}
    \item{\code{Cell_81}}{a numeric vector}
    \item{\code{Cell_82}}{a numeric vector}
    \item{\code{Cell_83}}{a numeric vector}
    \item{\code{Cell_84}}{a numeric vector}
    \item{\code{Cell_85}}{a numeric vector}
    \item{\code{Cell_86}}{a numeric vector}
    \item{\code{Cell_87}}{a numeric vector}
    \item{\code{Cell_88}}{a numeric vector}
    \item{\code{Cell_89}}{a numeric vector}
    \item{\code{Cell_9}}{a numeric vector}
    \item{\code{Cell_91}}{a numeric vector}
    \item{\code{Cell_92}}{a numeric vector}
    \item{\code{Cell_93}}{a numeric vector}
    \item{\code{Cell_94}}{a numeric vector}
    \item{\code{Cell_95}}{a numeric vector}
    \item{\code{Cell_96}}{a numeric vector}
    \item{\code{Cell_97}}{a numeric vector}
    \item{\code{Cell_98}}{a numeric vector}
    \item{\code{Cell_99}}{a numeric vector}
  }
}

\examples{
data(Ding_et_al._WBC1_TPMData)
## maybe str(Ding_et_al._WBC1_TPMData) ; plot(Ding_et_al._WBC1_TPMData) ...
}
\keyword{datasets}
