\name{gencode_v19_gene_pos}
\alias{gencode_v19_gene_pos}
\docType{data}
\title{genes position file}
\description{data frame contains Genes start and end position on chromosomes}
\usage{data("gencode_v19_gene_pos")}
\format{
  A data frame with 55765 observations on the following 3 variables.
  \describe{
  }
}

\examples{
data(gencode_v19_gene_pos)
## maybe str(gencode_v19_gene_pos) ; plot(gencode_v19_gene_pos) ...
}
\keyword{datasets}
