% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_Violin_plots.R
\name{Gene_Violin_plots}
\alias{Gene_Violin_plots}
\title{Gene Violin plot}
\usage{
Gene_Violin_plots(
  data_list = list(),
  data_id = list(),
  min_Sample = 5,
  min_Gene = 1500,
  gene_symbol = "",
  MetaData = list(),
  Groupby = "Clusters",
  DDLKCluster_data
)
}
\arguments{
\item{data_list}{List of expression matricies}

\item{data_id}{List of names/ids of expression matrix}

\item{min_Sample}{gene filter, filter out genes which are not expressed
in at least min_Sample cells}

\item{min_Gene}{cell filter, filter out those cells which do not
express at least min_Gene genes}

\item{gene_symbol}{Gene should present in all expression data of data_list}

\item{MetaData}{Optional, List of metadata of expression matricies in same
order in which expression matricies in data_list, Column number and names
of all the MetaData in the list must be same}

\item{Groupby}{column name in MetaData according
to which data is grouped.}

\item{DDLKCluster_data}{PathwayDDLK_clust from DDLK_Clust.R method}
}
\value{
Violin_plot violin plot and plot object of given gene.
}
\description{
Create Violin plot of specific gene
and compare expression in all data in data list.
}
\examples{
data1 = unCTC::Poonia_et_al._TPMData
data2 = unCTC::Ding_et_al._WBC1_TPMData
Data_list = list(data1,data2)
Data_Id = list("data1","data2")
Violin_plots = Gene_Violin_plots(data_list=Data_list,
                                 data_id= Data_Id,
                                 min_Sample = 5,
                                 min_Gene = 1500,
                                 Groupby = "data_id",
                                  gene_symbol ="TSPAN6" )
}
