% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stouffer_score.R
\name{get_stouffer}
\alias{get_stouffer}
\title{Calculate Stouffer's score}
\usage{
get_stouffer(data1, axis = 2)
}
\arguments{
\item{data1}{z score matrix}

\item{axis}{1 = applies over rows, 2 = applies over columns}
}
\value{
Stouffer's score
}
\description{
Stouffer's score computation
}
\examples{
data = unCTC::Poonia_et_al._TPMData
Filtered_data = data_filtering(data_mat= data,
                               min_sample =5,
                               min_gene=1500)
#Filtere data normalizationand log transformtion
library(Linnorm)
data_norm = log1p(Linnorm.Norm(Filtered_data))
data_zscore = get_Zscore(data_norm)
StoufferScore = get_stouffer(data_zscore,axis=2)

}
