% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stouffer_score.R
\name{get_Zscore}
\alias{get_Zscore}
\title{Claculate z-score of log transformed expression matrix}
\usage{
get_Zscore(data_mat)
}
\arguments{
\item{data_mat}{log transformed Expresion matrix,
genes in row and cells in column}
}
\value{
data_z zscore matrix
}
\description{
Claculate Z score of matrix
}
\examples{
data = unCTC::Poonia_et_al._TPMData
Filtered_data = data_filtering(data_mat= data,
                               min_sample =5,
                               min_gene=1500)
#Filtere data normalizationand log transformtion
library(Linnorm)
data_norm = log1p(Linnorm.Norm(Filtered_data))
data_zscore = get_Zscore(data_norm)

}
