% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupsDiffGenes.R
\name{GroupsDiffGenes}
\alias{GroupsDiffGenes}
\title{Claculate differential gene between two type of groups.}
\usage{
GroupsDiffGenes(
  data_mat,
  group,
  data_type = c("Normalised", "Raw"),
  Normalization_method = c("TMM", "TMMwsp", "RLE", "upperquartile", "none"),
  p_val = 0.05,
  lfc = 0,
  up_gene_number = 10
)
}
\arguments{
\item{data_mat}{Raw or Normalized data matrix in which genes in the row
and cells in columns.}

\item{group}{Different groups in a vector of size equals to the
sample size of data_mat}

\item{data_type}{whether data is Normalised or Raw (without normalization)}

\item{Normalization_method}{Only used if expresion matrix is not
normalized. All normalization methods are explained in calcNormFactor
function of edgeR package}

\item{p_val}{Threshold p value, default is 0.05}

\item{lfc}{Threshold log fold change value, default is 0}

\item{up_gene_number}{select number of upregulated genes in each group}
}
\value{
Up_gene_mat return expression matrix of upregulated genes
in each group
}
\description{
Calculate the top ten most differentially elevated genes
in group 1 that have a significant p-value difference.
For differentiation, limma voom is utilised.
}
\examples{
data = unCTC::Poonia_et_al._TPMData
groups = c(rep("TNBC",11),rep("NonTNBC",61))
output = GroupsDiffGenes(data_mat=data,
                          group=groups,
                          data_type="Raw",
                           Normalization_method="TMM")
}
