% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupsDiffPathways.R
\name{GroupsDiffPathways}
\alias{GroupsDiffPathways}
\title{Claculate differential pathways between two type of groups.}
\usage{
GroupsDiffPathways(
  Pathway_mat,
  group,
  p_val = 0.05,
  lfc = 0,
  up_pathways_number = 10
)
}
\arguments{
\item{Pathway_mat}{Pathway enrichment score matrix in which cells in the
columns and pathways in the rows.}

\item{group}{Different groups in a vector of size equals to the
sample size of Pathway_mat}

\item{p_val}{Threshold p value, default is 0.05}

\item{lfc}{Threshold log fold change value, default is 0}

\item{up_pathways_number}{select number of upregulated pathways
in each group}
}
\value{
Up_pathway_mat return expression matrix of 10 and 100 most
upregulated pathways respectively
}
\description{
Calculate top most differenciated upregulated pathways
with significant p-value between in group1, given in the form of vector.
Limma is used for differentiaition
}
\examples{
data = unCTC::Poonia_et_al._TPMData
genesets = unCTC::c2.all.v7.2.symbols
Pathways_mat = PathwayEnrichmentScore(data_list=list(data),
                                       data_id = list("data"),
                                       Genesets = genesets,
                                       min.size = 50,
                                       max.size = 100)
groups = c(rep("TNBC",11),rep("NonTNBC",61))
output = GroupsDiffPathways(Pathway_mat=Pathways_mat$Pathway_score,
                          group=groups)

}
