% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSingleCellObject.R
\name{Meta_data}
\alias{Meta_data}
\title{Meta Data computation}
\usage{
Meta_data(data_list = list(), data_id = list())
}
\arguments{
\item{data_list}{List of expression matricies}

\item{data_id}{List of names/ids of expression matricies}
}
\value{
batch_id1 return dataframe with data_id column
}
\description{
data frame with one column 'data_id'.
data_id is a vector of length equals to total number
of cells in all matricies.
}
\examples{
data1 = unCTC::Poonia_et_al._TPMData
data2 = unCTC::Ding_et_al._WBC1_TPMData
Data_list = list(data1,data2)
Data_Id = list("data1","data2")
metadata = Meta_data(data_list=Data_list,
                      data_id = Data_Id)

}
