% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayEnrichmentScore.R
\name{PathwayEnrichmentScore}
\alias{PathwayEnrichmentScore}
\title{Calculate Pathway enrichment score of all cells for all pathways}
\usage{
PathwayEnrichmentScore(
  data_list = list(),
  data_id = list(),
  Genesets,
  min.size = 10,
  max.size = 500,
  min_Sample = 5,
  min_Gene = 1500,
  Parallel_threads = 4L
)
}
\arguments{
\item{data_list}{List of gene expression data matricies. Genes/Features
should be in rows and cells/ samples in columns.}

\item{data_id}{List of names/ids of expression matrix}

\item{Genesets}{list of genesets/pathways.}

\item{min.size}{Minimum size of the resulting gene sets.}

\item{max.size}{Maximum size of the resulting gene sets.}

\item{min_Sample}{gene filter, filter out genes which are not expressed
in at least min_Sample cells}

\item{min_Gene}{cell filter, filter out those cells which do not
express at least min_Gene genes}

\item{Parallel_threads}{= Number of threads in parallel to execute process.}
}
\value{
PathwayData list of athway enrichment score and pathway metadata.
}
\description{
Calculate gene set enrichment across samples/cells using
the GSVA package in R, a non-parametric method and unsupervised
software programme.
}
\examples{
data1 = unCTC::Poonia_et_al._TPMData
data2 = unCTC::Ding_et_al._WBC1_TPMData
Data_list = list(data1,data2)
Data_Id = list("data1","data2")
Genesets = unCTC::c2.all.v7.2.symbols
Pathway_score = PathwayEnrichmentScore(data_list=Data_list,
                                        data_id= Data_Id,
                                        Genesets=Genesets,
                                        min.size=70,
                                        max.size=100,
                                        min_Sample = 5,
                                        min_Gene = 1500
                                        )
}
