\name{Poonia_et_al._CountData}
\alias{Poonia_et_al._CountData}
\docType{data}
\title{Poonia et al. CountData}
\description{Single cell RNA-seq count matrix of Circulating tumor cells
from poonia et al. data}
\usage{data("Poonia_et_al._CountData")}
\format{
  A data frame with 23732 observations on the following 72 variables.
  \describe{
    \item{\samp{1851009049_CS5_S22}}{a numeric vector}
    \item{\samp{1851013039_CS34_S53}}{a numeric vector}
    \item{\samp{1851009049_CS14_S17}}{a numeric vector}
    \item{\samp{1851009049_CS9_S25}}{a numeric vector}
    \item{\samp{1851009049_CS4_S21}}{a numeric vector}
    \item{\samp{1850055018_CS13_S2}}{a numeric vector}
    \item{\samp{1851013039_CS16_S33}}{a numeric vector}
    \item{\samp{1851013039_CS39_S58}}{a numeric vector}
    \item{\samp{1850061049_CS11_S1}}{a numeric vector}
    \item{\samp{1851009049_CS7_S23}}{a numeric vector}
    \item{\samp{1850055018_CS32_S12}}{a numeric vector}
    \item{\samp{1851013039_CS19_S36}}{a numeric vector}
    \item{\samp{1851013039_CS43_S63}}{a numeric vector}
    \item{\samp{1850061049_CS17_S3}}{a numeric vector}
    \item{\samp{1851013039_CS26_S44}}{a numeric vector}
    \item{\samp{1851030001-BC297CS2_S18}}{a numeric vector}
    \item{\samp{1851013039_CS11_S28}}{a numeric vector}
    \item{\samp{1851009049_CS11_S15}}{a numeric vector}
    \item{\samp{1850055018_CS24_S7}}{a numeric vector}
    \item{\samp{1851013039_CS1_S26}}{a numeric vector}
    \item{\samp{1851009049_CS17_S20}}{a numeric vector}
    \item{\samp{1850055018_CS18_S5}}{a numeric vector}
    \item{\samp{1851013039_CS30_S49}}{a numeric vector}
    \item{\samp{1851013039_CS15_S32}}{a numeric vector}
    \item{\samp{1851013039_CS6_S70}}{a numeric vector}
    \item{\samp{1851013039_CS32_S51}}{a numeric vector}
    \item{\samp{1851013039_CS21_S39}}{a numeric vector}
    \item{\samp{1850061049_CS24_S6}}{a numeric vector}
    \item{\samp{1851013039_CS40_S60}}{a numeric vector}
    \item{\samp{1851009049_CS16_S19}}{a numeric vector}
    \item{\samp{1850055018_CS25_S8}}{a numeric vector}
    \item{\samp{1851009049_CS10_S14}}{a numeric vector}
    \item{\samp{1851030113-BC130CS4_S12}}{a numeric vector}
    \item{\samp{1851013039_CS22_S40}}{a numeric vector}
    \item{\samp{1851013039_CS41_S61}}{a numeric vector}
    \item{\samp{1850061049_CS28_S8}}{a numeric vector}
    \item{\samp{1851013039_CS36_S55}}{a numeric vector}
    \item{\samp{1851013039_CS33_S52}}{a numeric vector}
    \item{\samp{1850055018_CS26_S9}}{a numeric vector}
    \item{\samp{1850061049_CS26_S7}}{a numeric vector}
    \item{\samp{1850055018_CS41_S15}}{a numeric vector}
    \item{\samp{1851013039_CS12_S29}}{a numeric vector}
    \item{\samp{1850061049_CS31_S11}}{a numeric vector}
    \item{\samp{1851013039_CS14_S31}}{a numeric vector}
    \item{\samp{1851009049_CS15_S18}}{a numeric vector}
    \item{\samp{1850061049_CS46_S12}}{a numeric vector}
    \item{\samp{1851030001-BC297CS4_S20}}{a numeric vector}
    \item{\samp{1851013039_CS25_S43}}{a numeric vector}
    \item{\samp{1851030113-BC130CS2_S10}}{a numeric vector}
    \item{\samp{1851013039_CS29_S47}}{a numeric vector}
    \item{\samp{1851013039_CS24_S42}}{a numeric vector}
    \item{\samp{1851013039_CS13_S30}}{a numeric vector}
    \item{\samp{1850061049_CS23_S5}}{a numeric vector}
    \item{\samp{1850055018_CS38_S14}}{a numeric vector}
    \item{\samp{1851013039_CS17_S34}}{a numeric vector}
    \item{\samp{1850055018_CS12_S1}}{a numeric vector}
    \item{\samp{1851013039_CS3_S48}}{a numeric vector}
    \item{\samp{1850055018_CS31_S11}}{a numeric vector}
    \item{\samp{1851030113-BC130CS3_S11}}{a numeric vector}
    \item{\samp{1850061049_CS12_S2}}{a numeric vector}
    \item{\samp{1851013039_CS37_S56}}{a numeric vector}
    \item{\samp{1851009049_CS8_S24}}{a numeric vector}
    \item{\samp{1851013039_CS23_S41}}{a numeric vector}
    \item{\samp{1851030001-BC297CS7_S23}}{a numeric vector}
    \item{\samp{1850055018_CS17_S4}}{a numeric vector}
    \item{\samp{1851013039_CS2_S37}}{a numeric vector}
    \item{\samp{1851013039_CS31_S50}}{a numeric vector}
    \item{\samp{1851013039_CS47_S67}}{a numeric vector}
    \item{\samp{1851030113-BC130CS5_S13}}{a numeric vector}
    \item{\samp{1851013039_CS18_S35}}{a numeric vector}
    \item{\samp{1850061049_CS29_S9}}{a numeric vector}
    \item{\samp{1851009049_CS12_S16}}{a numeric vector}
  }
}

\examples{
data(Poonia_et_al._CountData)
## maybe str(Poonia_et_al._CountData) ; plot(Poonia_et_al._CountData) ...
}
\keyword{datasets}
