\name{Poonia_et_al._metaData}
\alias{Poonia_et_al._metaData}
\docType{data}
\title{Poonia et al. metaData}
\description{Meta data file of Circulating Tumor Cells from Poonia et al.}
\usage{data("Poonia_et_al._metaData")}
\format{
  A data frame with 72 observations on the following 4 variables.
  \describe{
    \item{\code{GroupID}}{a character vector}
    \item{\code{Cell_type}}{a character vector}
    \item{\code{HormoneStatus}}{a character vector}
    \item{\code{Class}}{a character vector}
  }
}
\examples{
data(Poonia_et_al._metaData)
## maybe str(Poonia_et_al._metaData) ; plot(Poonia_et_al._metaData) ...
}
\keyword{datasets}
