\name{Poonia_et_al._PBMC_CountData}
\alias{Poonia_et_al._PBMC_CountData}
\docType{data}
\title{Poonia et al. WBC Count Data}
\description{Single cell RNA-seq count matrix of PBMCs
from poonia et al. data}
\usage{data("Poonia_et_al._PBMC_CountData")}
\format{
  A data frame with 23732 observations on the following 71 variables.
  \describe{
    \item{\samp{C1-022-C14_S26}}{a numeric vector}
    \item{\samp{C1-022-C41_S80}}{a numeric vector}
    \item{\samp{C1-022-C75_S52}}{a numeric vector}
    \item{\samp{C1-022-C77_S59}}{a numeric vector}
    \item{\samp{C1-022-C43_S85}}{a numeric vector}
    \item{\samp{C1-022-C80_S65}}{a numeric vector}
    \item{\samp{C1-022-C02_S2}}{a numeric vector}
    \item{\samp{C1-022-C33_S63}}{a numeric vector}
    \item{\samp{C1-022-C70_S46}}{a numeric vector}
    \item{\samp{C1-022-C01_S3}}{a numeric vector}
    \item{\samp{C1-022-C32_S62}}{a numeric vector}
    \item{\samp{C1-022-C10_S21}}{a numeric vector}
    \item{\samp{C1-022-C22_S45}}{a numeric vector}
    \item{\samp{C1-022-C51_S6}}{a numeric vector}
    \item{\samp{C1-022-C47_S92}}{a numeric vector}
    \item{\samp{C1-022-C96_S94}}{a numeric vector}
    \item{\samp{C1-022-C91_S90}}{a numeric vector}
    \item{\samp{C1-022-C24_S43}}{a numeric vector}
    \item{\samp{C1-022-C46_S91}}{a numeric vector}
    \item{\samp{C1-022-C54_S12}}{a numeric vector}
    \item{\samp{C1-022-C61_S28}}{a numeric vector}
    \item{\samp{C1-022-C07_S15}}{a numeric vector}
    \item{\samp{C1-022-C30_S57}}{a numeric vector}
    \item{\samp{C1-022-C04_S9}}{a numeric vector}
    \item{\samp{C1-022-C95_S95}}{a numeric vector}
    \item{\samp{C1-022-C63_S30}}{a numeric vector}
    \item{\samp{C1-022-C09_S13}}{a numeric vector}
    \item{\samp{C1-022-C81_S64}}{a numeric vector}
    \item{\samp{C1-022-C37_S73}}{a numeric vector}
    \item{\samp{C1-022-C17_S32}}{a numeric vector}
    \item{\samp{C1-022-C84_S70}}{a numeric vector}
    \item{\samp{C1-022-C48_S93}}{a numeric vector}
    \item{\samp{C1-022-C18_S31}}{a numeric vector}
    \item{\samp{C1-022-C34_S67}}{a numeric vector}
    \item{\samp{C1-022-C21_S37}}{a numeric vector}
    \item{\samp{C1-022-C44_S86}}{a numeric vector}
    \item{\samp{C1-022-C92_S89}}{a numeric vector}
    \item{\samp{C1-022-C58_S22}}{a numeric vector}
    \item{\samp{C1-022-C69_S42}}{a numeric vector}
    \item{\samp{C1-022-C11_S20}}{a numeric vector}
    \item{\samp{C1-022-C82_S72}}{a numeric vector}
    \item{\samp{C1-022-C66_S36}}{a numeric vector}
    \item{\samp{C1-022-C79_S66}}{a numeric vector}
    \item{\samp{C1-022-C08_S14}}{a numeric vector}
    \item{\samp{C1-022-C40_S79}}{a numeric vector}
    \item{\samp{C1-022-C26_S50}}{a numeric vector}
    \item{\samp{C1-022-C12_S19}}{a numeric vector}
    \item{\samp{C1-022-C52_S10}}{a numeric vector}
    \item{\samp{C1-022-C50_S5}}{a numeric vector}
    \item{\samp{C1-022-C86_S77}}{a numeric vector}
    \item{\samp{C1-022-C35_S68}}{a numeric vector}
    \item{\samp{C1-022-C73_S54}}{a numeric vector}
    \item{\samp{C1-022-C27_S51}}{a numeric vector}
    \item{\samp{C1-022-C05_S8}}{a numeric vector}
    \item{\samp{C1-022-C85_S78}}{a numeric vector}
    \item{\samp{C1-022-C68_S41}}{a numeric vector}
    \item{\samp{C1-022-C16_S33}}{a numeric vector}
    \item{\samp{C1-022-C76_S60}}{a numeric vector}
    \item{\samp{C1-022-C62_S29}}{a numeric vector}
    \item{\samp{C1-022-C53_S11}}{a numeric vector}
    \item{\samp{C1-022-C25_S49}}{a numeric vector}
    \item{\samp{C1-022-C71_S47}}{a numeric vector}
    \item{\samp{C1-022-C60_S24}}{a numeric vector}
    \item{\samp{C1-022-C65_S35}}{a numeric vector}
    \item{\samp{C1-022-C64_S34}}{a numeric vector}
    \item{\samp{C1-022-C83_S71}}{a numeric vector}
    \item{\samp{C1-022-C13_S27}}{a numeric vector}
    \item{\samp{C1-022-C06_S7}}{a numeric vector}
    \item{\samp{C1-022-C31_S61}}{a numeric vector}
    \item{\samp{C1-022-C28_S55}}{a numeric vector}
    \item{\samp{C1-022-C57_S18}}{a numeric vector}
  }
}

\examples{
data(Poonia_et_al._PBMC_CountData)
## maybe str(Poonia_et_al._PBMC_CountData) ; plot(Poonia_et_al._PBMC_CountData) ...
}
\keyword{datasets}
