\name{Poonia_et_al._PBMC_metaData}
\alias{Poonia_et_al._PBMC_metaData}
\docType{data}
\title{Poonia et al. WBC metaData}
\description{Meta data file of  single cell RNA-seq PBMC data
from Poonia et al.}
\usage{data("Poonia_et_al._PBMC_metaData")}
\format{
  A data frame with 71 observations on the following 4 variables.
  \describe{
    \item{\code{GroupID}}{a character vector}
    \item{\code{Cell_type}}{a character vector}
    \item{\code{HormoneStatus}}{a character vector}
    \item{\code{Class}}{a character vector}
  }
}

\examples{
data(Poonia_et_al._PBMC_metaData)
## maybe str(Poonia_et_al._PBMC_metaData) ; plot(Poonia_et_al._PBMC_metaData) ...
}
\keyword{datasets}
