% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stouffer_score.R
\name{Stouffer_score}
\alias{Stouffer_score}
\title{Calculate Stouffer's Score for specific list of genes and plot
stouffer's score between different type of cells.}
\usage{
Stouffer_score(
  data_list = list(),
  min_Sample = 5,
  min_Gene = 1500,
  gene_list = list(),
  data_id = list(),
  MetaData = list(),
  metaColPos = 1,
  metaColName = "Class",
  Groupby = "Clusters",
  DDLKCluster_data
)
}
\arguments{
\item{data_list}{List of expression matrix}

\item{min_Sample}{gene filter, filter out genes which are not expressed
in at least min_Sample cells}

\item{min_Gene}{cell filter, filter out those cells which do not
express at least min_Gene genes}

\item{gene_list}{Vector of specific genes for which we
want to calculate Stouffer's score}

\item{data_id}{List of names/ids of expression matrix}

\item{MetaData}{Optional, List of metadata of expression matricies in same
order in which expression matricies in data_list, Column number and names
of all the MetaData in the list must be same}

\item{metaColPos}{column position in MetaData according
to which we want to group data, default is 1}

\item{metaColName}{Name of the column in MetaData according
to which we want to group data, default is "Class"}

\item{Groupby}{column name in MetaData according
to which data is grouped.}

\item{DDLKCluster_data}{output of DDLK_Clust.R method}
}
\value{
genes_Stouffer_score1 Cell wise stouffer score with Class info
}
\description{
Calculate Stouffer,s score of genes passed as gene_list in
Stouffer_score()
}
\examples{
data1 = unCTC::Poonia_et_al._TPMData
data2 = unCTC::Ding_et_al._WBC1_TPMData
Data_list = list(data1,data2)
Data_Id = list("data1","data2")
#LincrNA specific gene list
gene_list = unCTC::Breast_elevated_genes
S = Stouffer_score(data_list=Data_list,
                    min_Sample = 5,
                    min_Gene = 1500,
                    gene_list = gene_list,
                    Groupby = "data_id",
                    data_id= Data_Id)

}
