% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unCTC_plots.R
\name{unCTC_plots}
\alias{unCTC_plots}
\title{Pathway based principal component analysis plots}
\usage{
unCTC_plots(
  Pathway_score,
  Pathway_metadata,
  colorby = "Data_id",
  Color_cluster = "Clusters",
  pairsplotLegend = c("left", "right", "none")
)
}
\arguments{
\item{Pathway_score}{Pathway score from DDLK_Clust function}

\item{Pathway_metadata}{PathwayDDLK_clust from DDLK_Clust function}

\item{colorby}{Any column name from Pathway_metadata,
default is "Data_id"}

\item{Color_cluster}{Any column name from Pathway_metadata,
default is "Clusters"}

\item{pairsplotLegend}{Legend position for pairsplot.}
}
\value{
plots list of color by class, color by clusters, pairsplot
}
\description{
Plot unCTC results
}
\examples{
data1 = unCTC::Poonia_et_al._TPMData
data2 = unCTC::Ding_et_al._WBC1_TPMData
Data_list = list(data1,data2)
Data_Id = list("data1","data2")
Genesets = unCTC::c2.all.v7.2.symbols
Pathway_score = PathwayEnrichmentScore(data_list=Data_list,
                                        data_id= Data_Id,
                                        Genesets=Genesets,
                                        min.size=70,
                                        max.size=100)

DDLK_Clusters = DDLK_Clust(PathwayScore = Pathway_score$Pathway_score,
                           PathwayMetaData=Pathway_score$Pathway_metadata,
                            n=3,
                            out.dir = getwd())

PathwayScore = DDLK_Clusters$Pathway_score
PathwayMetadata = DDLK_Clusters$PathwayDDLK_clust
Plots_output = unCTC_plots(Pathway_score=PathwayScore,
                           Pathway_metadata = PathwayMetadata,
                           colorby = "Data_id",
                           Color_cluster = "Clusters")
}
